requirements.txt
Flask==3.0.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
bcrypt==4.1.2 
pillow==10.2.0
email-validator==2.1.0.post1
flask_security ==3.0.0